Whitespace Normalization of Google TTS Text Input
In the use of psuedo "tts" tags in defining the speech text on various
cards on various templates, there can be differences in whitespace
between what are otherwise the same expressions.

This adds an additional regular expression replacement to the text input
before hashing it, collapsing all whitespace down and trimming it,
decreasing the number of cache misses.