Exception Handling for Bad TTS Calls
This adds on-screen error messages for when a urllib2 urlopen() call
either throws an exception or the response back from Google TTS is not
what is expected (i.e. an audio/mpeg file with a status of 200).

Both types of error messages will only be displayed once, which is
controlled by two different module-level flags.