Caching Direction Definition in Config Module
Moves the caching directory initialization to the config module, so that
it may be more accessible elsewhere in the program (notably, the code
that initializes the configuration screen).

Also, in the future, it might be nice to let the user specify the
directory to use for caching (and in that way, the user could cache to a
network or shared drive if s/he wanted to share the cache among multiple
instances of the program).