Exception when Emptying Cache w/ Files In-Use
When emptying the cache from the AwesomeTTS configuration screen, if a
file cannot be deleted for some reason, it's tracked and the message on
the button is updated appropriately once the entire operation is
completed explaining how many files could not be deleted.

A common cause of this is on the Windows operating system, where if a
long audio clip has been played back and "enable subprocessing" is
checked in the configuration screen, the mplayer process dealing with it
may not exit correctly until Anki itself is closed. In this case, the
MP3 files the process was playing when it got stuck will be "in use" and
will not be clearable by the user.