Move Google TTS Download to Thread Class
Takes the current downloader logic and moves it to a "worker" class
which is accessed by a "downloader" class. The latter is in charge of
firing up the thread and making sure that the user does not do anything
that would cause another thread for the same file to be started.