Install/symlink fix for systems with BSD mktemp
When using mktemp on a system with BSD userland utilities (like Mac OS X
has), a "template" must be provided (i.e. a vague desired filename for
the temporary file, e.g. /tmp/conf.db.XXXXXXXXXX). Linux works with or
without the "template", and is not harmed by its presence.