Install/symlink check for parent directory
When running the install or symlink scripts, care should be taken to
ensure that the target directory is not a parent directory of the
package itself.

This can happen accidentally if one already has a symlink from the
addons directory to the project directory, changes into the symlinked
project directory, forgets about being in the symlink, and then asks for
a symlink/install again back to the addon directory.