Bind keyPressEvent to dialog in configuration
When attempting to assign keyboard shortcuts for TTS playback, Mac OS X
is unable to capture the key press event if bound to the buttons. (The
buttons are unfocused after being clicked on that platform, so the key
press event might not be able to reach the buttons.)

Instead, we can try to bind the key press event to the dialog box
itself, which seems to always be able to catch the event. In this
variation, we can also check to see if both buttons are awaiting the
event, and assign both if so (now that using the same keyboard shortcut
on both the fronts and backs of cards is supported).