Improve speed of Google record() when cached
During the record() hook, first check to see if we already have the
image locally in the cache directory. If so, pass that path to Anki for
inclusion in the media collection.

Additionally, makes adjustments to various places in the main module to
prevent removal of the cached asset and do not apply throttling when
retrieving cached assets.