Buttons and move the delayed UI construction
Introduces buttons into the UI mix.

Unfortunately, the delayed construction code needed to be moved and
reworked. The Qt dialog/modal/parent functionality only works properly
when construction happens with the parent already known, so we cannot
keep a form in memory once it's loaded.

Instead, we can defer construction of a form until it's clicked, and
that is nice, but if the button or menu action that is responsible for
triggering the form goes away, then the form needs to be disposed of.

We can, however, maintain the same form in memory as long as the button
or menu action responsible for triggering it stays alive.