Fix for missing arguments in signals
This was happening because the Worker was being deleted before its
signal was passed. This has been fixed now so that the delete is not
permitted until the signal is passed and the callback is executed.

(cherry picked from commit 8c8cee620a1f56300f5aaae12e296cb40ef8a446)