Fix for shortcut conflict on Browser window
While in the Browser window's Editor mode, two shortcuts were active
simultaneously (i.e. the Browser Generator menu shortcut in the Edit
menu and the more local Editor Generator button along the toolbar),
which means that only the Browser Generator was usable. Given the
context, the user probably actually wants the Editor Generator.

This adjusts the action menu shortcut to become disabled when focus is
gained on the Editor field, and re-enabled when focus is lost, allowing
the more "local" shortcut to prevail.