Fix for Finder indexing caches on Mac OS X
Because Anki puts its directory into the user's Documents directory on
Mac OS X and because Mac OS X attempts to be helpful by displaying all
MP3s in the user's Documents directory when opening Finder, it was
discovering and displaying all cached MP3s as if they were music.

This renames the cache and temp directories to have dot prefixes so Mac
OS X will get the hint to leave them alone.