Improve error message when missing lame
When install Anki on Ubuntu, for example, lame and mplayer are not
automatically installed. Since we use Anki's interface for audio
playback, the missing mplayer issue is taken care of, but we use lame
directly.

This adds a try/except around the lame CLI call to catch if the call
fails because the binary is missing, and raises a better message.