Remove normalization functionality from router
Because the router does not have full context of where text is coming
from (e.g. note field, on-the-fly text, user input), it really doesn't
have all the wherewithal to figure out the best way to normalize it.

This becomes more problematic when we start dealing with how to treat
cloze and parenthetical text in a sane manner. With this branch, it will
become the responsibility of the caller to the router to figure out how
to do normalizations of the text.