Replace STRIP_ALL with more nuanced stripping
This adds new stripping options in the 'addon' bundle to include
strip.from_note (for BrowserGenerator runs, future prepopulation of
modals from note fields), strip.from_template (for on-the-fly),
strip.from_user (for direct user input in previews or EditorGenerator
insertion), and strip.sounds (to continue to allow stripping of sounds
where necessary, e.g. the BrowserGenerator's removal of old sounds or
the Reviewer's handling of faking the {{FrontSide}} placeholder that
Anki uses).