Reviewer regression with "empty" text
The reviewer (i.e. on-the-fly mode) is supposed to silently ignore empty
tags, but the stripping was placed in the wrong place causes a
regression in that guarantee.

This fixes that by moving the text stripping to be immediately above the
'if not text' check that does the 'continue' to ignore the tag.