Shortcut configuration fix for Mac OS X
With the introduction of text-like input fields on the first tab of the
configuration window, Mac OS X's implementation of Qt now cannot capture
keystrokes from the shortcut buttons.

This fix makes it so that a focus event is automatically fired on the
event buttons to focus them explicitly so that the keyboard events can
be captured.