Exit text transformation early, if possible
If during a text transformation, we are left with an empty string, exit
out instead of continuing to process the rulesets.

This can be helpful in cases where we might try multiple things to get a
text string (like when launching the note editor), and we do not need to
go through every rule when we have nothing to work with.