Bump grunt-contrib-htmlmin again
Updates `grunt-contrib-htmlmin` now to v0.6.0, which brings the
underlying `htmlmin` library to v1.0.0.

This bump actually has an effect on the web build: links to the home
page are no longer quoted, as such attribute values cannot be confused
with self-closing tag if they are not the last attribute value in the
tag.