Suggest service groups when dictionaries fail
Marks Duden, Howjsay, and Oxford as "dictionary" services, and give the
router a method for determining if a particular service is of some
trait.

Finally, teaches the reviewer to tell the user about service groups when
they are playing audio on-the-fly for a single service but that single
service does not have audio.

Fixes #97.